python-api/requirements.txt
appdynamics==25.1.0.7425
appdynamics-bindeps-linux-x64==24.12.0
appdynamics-proxysupport-linux-x64==11.72.3
asgiref==3.8.1
azure-core==1.33.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.21.0
azure-monitor-opentelemetry==1.6.8
azure-monitor-opentelemetry-exporter==1.0.0b36
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
fixedint==0.1.6
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.2.2
packaging==25.0
psutil==6.1.1
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
pyuwsgi==2.0.28.post1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
